+++ src/spectrum/modcovar.py
@@ -297,10 +297,11 @@ class pmodcovar(ParametricSpectrum):
 
         For a detailled description of the parameters, see :func:`modcovar`.
 
-        :param array data:
+        :param array data:     input data (list or numpy.array)
         :param int order:  
-        :param int NFFT:
-        :param float sampling: 
+        :param int NFFT:       total length of the final data sets (padded with zero if needed; default is 4096)
+        :param float sampling: sampling frequency of the input :attr:`data`.
+
 
         """
         super(pmodcovar, self).__init__(data, ar_order=order, 
@@ -311,7 +312,7 @@ class pmodcovar(ParametricSpectrum):
         from spectrum import arma2psd
         ar, e = modcovar(self.data, self.ar_order)
         self.ar = ar
-        psd = arma2psd(A=ar, T=self.sampling, NPSD=self.NFFT)
+        psd = arma2psd(A=ar, T=self.sampling, NFFT=self.NFFT)
         
         if self.datatype == 'real':
             from tools import twosided_2_onesided