+++ test/test_spectrum.py
@@ -4,4 +4,4 @@ from spectrum import *
 
 def test_spectrum():
     assert spectrum.__version__
-    assert spectrum.default_NPSD
+    assert spectrum.default_NFFT